Move add-project-form into ProjectForm component
